refactor(controller): extract reconciler::mcp_egress submodule
Move the MCP egress derivation (mcp_url_host_port / cluster_internal_namespace
/ mcp_egress_rule + the per-ref derive_mcp_egress_rules loop) and its unit
tests out of reconciler/mod.rs into a focused reconciler::mcp_egress module.
Keeps mod.rs under the ci/check-loc.sh phase0 budget (3700 LOC) and groups the
egress logic in one place. Pure code-move + the reconcile loop now calls
derive_mcp_egress_rules; no behaviour change. 854 controller tests pass.
